UnBArqDsw/2020.1_G13_iGado#113 📦 NEW: Create report utils, separate all classes and apply design patterns
