+++ project/api/models/farm.py
@@ -11,7 +11,7 @@ class FarmModel(db.Model):
     size_farm = db.Column(db.Integer, nullable=False)
     users = relationship('UserModel', secondary='work')
     generalReports = relationship('GeneralReportModel', backref='farm', lazy=True)
-    gmdReports = relationship('ReportGMDModel', backref='farm', lazy=True)
+    gmdReports = relationship('GMDReportModel', backref='farm', lazy=True)
     # beef_cattles = relationship('BeefCattleModel', backref='farm', lazy=True)
     # dairy_cattles = relationship('DairyCattleModel', backref='farm', lazy=True)
